Levrage Dt args in data nft factory.
